+++ folding-stats-jar/src/main/java/me/zodac/folding/db/postgres/RecordConverter.java
@@ -53,7 +53,7 @@ final class RecordConverter {
             HardwareMake.get(hardwareRecord.getHardwareMake()),
             HardwareType.get(hardwareRecord.getHardwareType()),
             hardwareRecord.getMultiplier().doubleValue(),
-            hardwareRecord.getAveragePpd().doubleValue()
+            hardwareRecord.getAveragePpd().longValue()
         );
     }
